+++ assets/js/scripts-global.js
@@ -5,7 +5,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0
- * @version  2.0
+ * @version  2.1.1
  *
  * Contents:
  *
@@ -215,7 +215,7 @@ jQuery( function() {
 							// Processing
 
 								$postsContainers.masonry( {
-										itemSelector : '.hentry'
+										itemSelector : '.entry'
 									} );
 
 						} );